+++ mettagrid/src/metta/mettagrid/mettagrid_c.hpp
@@ -34,8 +34,20 @@ class ConverterConfig;
 class WallConfig;
 class AgentConfig;
 class GameConfig;
+class ActionConfig;
 
 namespace py = pybind11;
+
+struct GameConfig {
+  int num_agents;
+  unsigned int max_steps;
+  unsigned short obs_width;
+  unsigned short obs_height;
+  std::vector<std::string> inventory_item_names;
+  unsigned int num_observation_tokens;
+  std::map<std::string, std::shared_ptr<ActionConfig>> actions;
+  std::map<std::string, std::shared_ptr<GridObjectConfig>> objects;
+};
 
 class METTAGRID_API MettaGrid {
 public:
